:pencil: Renamed all CSS modules to include .module extension
