+++ packages/mechanic-ui-components/src/buttons/Toggle.js
@@ -2,7 +2,7 @@ import React, { useState, useRef } from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
 import { uid } from "../uid.js";
-import * as css from "./Toggle.css";
+import * as css from "./Toggle.module.css";
 
 export const Toggle = props => {
   const _id = useRef(uid("toggle"));
