+++ packages/mechanic-ui-components/src/input/ColorInput.module.css
