+++ packages/mechanic-ui-components/src/input/NumberInput.module.css
