+++ packages/mechanic-ui-components/src/input/OptionInput.module.css
