+++ packages/mechanic-ui-components/src/input/Select.js
@@ -2,7 +2,7 @@ import React, { Fragment, useRef, useState } from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
 import { uid } from "../uid.js";
-import * as css from "./Select.css";
+import * as css from "./Select.module.css";
 
 export const Select = props => {
   const _id = useRef(uid("select"));
