+++ packages/mechanic/app/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import { App } from "./App.js";
-import "./index.css";
+import "./index.module.css";
 
 const root = document.getElementById("root");
 const app = (
