+++ packages/mechanic/app/index.module.css
@@ -12,7 +12,7 @@ html {
 body {
   padding: 0;
   margin: 0;
-  
+
   font-size: 14px;
   font-family: Helvetica, sans-serif;
   -webkit-font-smoothing: antialiased;
