+++ packages/mechanic/app/pages/NotFound.js
@@ -1,5 +1,5 @@
 import React from "react";
-import * as css from "./NotFound.css";
+import * as css from "./NotFound.module.css";
 
 export const NotFound = props => {
   return (
